requirements.txt
python-dotenv

# Inspect AI framework (UK AISI)
inspect-ai>=0.3.0

# LLM API dependencies
openai
anthropic
google-genai

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development and utility dependencies
typing-extensions>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
ruff>=0.1.0 